+++ MeowdoptMeApp/data/ads/Beza.json
@@ -1,6 +1,7 @@
 {
   "pet": {
     "name": "Beza",
+    "about": "Beza to pies, który trafił do nas wychudzony i schorowany, jednak dzięki właściwej opiece, stanęła na nogi. Jest wulkanem energii, świetnie chodzi na smyczy, uwielbia kontakt z ludźmi. Najlepiej odnajdzie się w domu z ogrodem.",
     "petCharacteristics": {
       "species": "dog",
       "breed": "Labrador Retriever mix",
@@ -17,10 +18,11 @@
     "photos": [
       {
         "img": "26",
-        "description": "Beza to pies, który trafił do nas wychudzony i schorowany, jednak dzięki właściwej opiece, stanęła na nogi. Jest wulkanem energii, świetnie chodzi na smyczy, uwielbia kontakt z ludźmi. Najlepiej odnajdzie się w domu z ogrodem."
+        "description": "beza na wyrze"
       },
       {
-        "img": "27"
+        "img": "27",
+        "description": "papuez"
       },
       {
         "img": "28"
