+++ MeowdoptMeApp/data/ads/Drops.json
@@ -1,6 +1,7 @@
 {
   "pet": {
     "name": "Drops",
+    "about": "Dropsik trafił do nas stosunkowo niedawno, w związku z przeprowadzką jego właściciela. Jest przyzwyczajony do stałej obecności człowieka. Bardzo dobrze chodzi na smyczy. Pilnie szukamy dla Dropsa domu, który zapewni mu dobre warunki i domowe ciepło.",
     "petCharacteristics": {
       "species": "dog",
       "breed": "mixed breed",
@@ -16,8 +17,7 @@
     "id": 3,
     "photos": [
       {
-        "img": "36",
-        "description": "Dropsik trafił do nas stosunkowo niedawno, w związku z przeprowadzką jego właściciela. Jest przyzwyczajony do stałej obecności człowieka. Bardzo dobrze chodzi na smyczy. Pilnie szukamy dla Dropsa domu, który zapewni mu dobre warunki i domowe ciepło."
+        "img": "36"
       },
       {
         "img": "37"
