+++ MeowdoptMeApp/data/ads/Mariusz.json
@@ -1,6 +1,7 @@
 {
   "pet": {
     "name": "Mariusz",
+    "about": "Mariusz to kot, który trafił do schroniska po śmierci swojego właściciela. Uwielbia kontakt i bliskość z człowiekiem. Umie korzystać z kuwety. W schronisku czuje się przytłoczony. Z niecierpliwością czeka na odpowiedzialny dom pełen miłości.",
     "petCharacteristics": {
       "species": "cat",
       "breed": "mixed breed",
@@ -16,8 +17,7 @@
     "id": 5,
     "photos": [
       {
-        "img": "06",
-        "description": "Mariusz to kot, który trafił do schroniska po śmierci swojego właściciela. Uwielbia kontakt i bliskość z człowiekiem. Umie korzystać z kuwety. W schronisku czuje się przytłoczony. Z niecierpliwością czeka na odpowiedzialny dom pełen miłości."
+        "img": "06"
       },
       {
         "img": "07"
