+++ MeowdoptMeApp/data/ads/Zuzia.json
@@ -1,6 +1,7 @@
 {
   "pet": {
     "name": "Zuzia",
+    "about": "Zuzia to pies, który prawdopodobnie uciekł z poprzedniego. Jednak w związku z brakiem odpowiedzi na ogłoszenia o jej znalezeniu oraz brakiem chipa identyfikującego, szukamy dla Zuzi nowego, odpowiedzialnego domu. Krótki pobyt w schronisku nie zdążył wpłynąć na nią negatywnie. Jest psem typowo kanapowym, wymaga nauki chodzenia na smyczy.",
     "petCharacteristics": {
       "species": "dog",
       "breed": "mixed breed",
@@ -16,8 +17,7 @@
     "id": 10,
     "photos": [
       {
-        "img": "21",
-        "description": "Zuzia to pies, który prawdopodobnie uciekł z poprzedniego. Jednak w związku z brakiem odpowiedzi na ogłoszenia o jej znalezeniu oraz brakiem chipa identyfikującego, szukamy dla Zuzi nowego, odpowiedzialnego domu. Krótki pobyt w schronisku nie zdążył wpłynąć na nią negatywnie. Jest psem typowo kanapowym, wymaga nauki chodzenia na smyczy."
+        "img": "21"
       },
       {
         "img": "22"
