+++ MeowdoptMeApp/data/ads/Łasuch.json
@@ -1,6 +1,7 @@
 {
   "pet": {
     "name": "Łasuch",
+    "about": "Łasuch to kot znaleziony w kartonach pod sklepem. Początkowo nieufny w stosunku do ludzi, jednak coraz bardziej przyzwyczaja się do dotyku i głaskania. Jego stan zdrowia jest bardzo dobry, ma jedynie czasem problem z łzawiącymi oczami. Łasuch czeka na dobry i kochający dom.",
     "petCharacteristics": {
       "species": "cat",
       "breed": "mixed breed",
@@ -16,8 +17,7 @@
     "id": 4,
     "photos": [
       {
-        "img": "01",
-        "description": "Łasuch to kot znaleziony w kartonach pod sklepem. Początkowo nieufny w stosunku do ludzi, jednak coraz bardziej przyzwyczaja się do dotyku i głaskania. Jego stan zdrowia jest bardzo dobry, ma jedynie czasem problem z łzawiącymi oczami. Łasuch czeka na dobry i kochający dom."
+        "img": "01"
       },
       {
         "img": "02"
