+++ MeowdoptMeApp/src/commonTypes.ts
@@ -9,22 +9,17 @@ export interface Ad {
 export interface PhotoAlbum {
   id: number;
   photos: Photo[];
-  // instead of add
-  // addPhoto: (photo: Photo) => void;
-  // instead of delete
-  // removePhoto: (photo: Photo) => void;
 }
 
 export interface Pet {
   name: string;
+  about: string;
   petCharacteristics: PetCharacteristics;
 }
 
 export interface Photo {
   img: string;
   description: string;
-  // setImg: (img: object) => void;
-  // setDescription: (description: string) => void;
 }
 
 export interface PetCharacteristics {
